Add JSON and fetch vocabs from GitHub
- Create README
