+++ scripts/vocabs_csv2jsonl/README.md
@@ -1,0 +1,83 @@
+# vocabs_csv2jsonl.py — Convert vocabulary CSV files to JSONL or JSON
+
+## Purpose
+Convert one or more vocabulary CSV files into machine-friendly JSONL or JSON files used by the application.  
+Features:
+- Optionally fetch CSVs from a remote raw GitHub URL, with a local fallback.
+- Produce a metadata header describing the source, generation time, format and count.
+
+---
+
+## Usage (CLI)
+```bash
+python3 vocabs_csv2jsonl.py [--local] [--remote-base URL] [--format jsonl|json] [--dst DEST]
+```
+
+### Options
+- `--remote-base`  
+    Base raw URL where CSVs live (default points to a project GitHub raw path).  
+    If provided and accessible, CSVs are fetched from `{remote_base}/{vocab}.csv`.
+- `--local`  
+    Force using local CSV files from `docs/vocabularies` instead of fetching remotely. When set, `--remote-base` is ignored.
+- `--format`  
+    Output format: `jsonl` (default) or `json`.  
+    - `jsonl`: first line is a metadata JSON object, followed by one JSON entry per line.  
+    - `json`: a single JSON object containing metadata and an `entries` array.
+- `--dst`  
+    Destination folder for generated files (default: `react-app/public/data`).
+
+---
+
+## Behavior and rules
+- Vocabularies processed are defined in the `VOCABS` mapping (e.g., `access_rights`, `licenses`, `media_types`, ...).
+- Remote mode: the script attempts to fetch each CSV; on failure it falls back to the local file if available.
+- CSV reading:
+    - Files opened with UTF-8 (`utf-8-sig`) and newline handling.
+    - Lines beginning with `#` (after left-strip) and empty lines are ignored.
+    - If the first non-comment line contains `://`, the file is treated as headerless data (each row parsed as CSV columns `col0`, `col1`, ...).
+    - Otherwise the first non-comment line is treated as a header row and parsed with `csv.DictReader`.
+- Normalization rules:
+    - `media_types`: first non-empty cell is used as `uri`.
+    - `licenses`: recognizes columns like `authority_uri`/`authority`, `code`/`id`, `url`/`license_url`; output may include `uri`, `code`, `url` and a `label` (from `code` or a slug from `uri`).
+    - Default (`uri`/`label` style): looks for `uri`/`url`/`col0` and `label`/`name`/`col1`. Missing labels are generated by slugging the URI (strip trailing `/` or `#`, take last segment, replace `_` and `-` with spaces).
+    - Rows missing required data (e.g., no URI) are skipped.
+
+---
+
+## Output metadata
+Each output file includes a metadata object with keys:
+- `source` — origin of the CSV
+- `generated` — UTC ISO8601 timestamp
+- `format` — `jsonl` or `json`
+- `type` — vocabulary type
+- `name` — vocabulary name
+- `count` — number of entries
+
+---
+
+## Error handling
+- Network or parsing errors print warnings to `stderr` and skip problematic sources.
+- If a remote CSV yields zero entries, the script will try a local fallback CSV for that vocabulary.
+
+---
+
+## Examples
+- Use remote CSVs and produce `jsonl` in default destination:
+```bash
+python3 vocabs_csv2jsonl.py
+```
+- Use local CSVs and produce a single JSON file:
+```bash
+python3 vocabs_csv2jsonl.py --local --format json --dst ./out/data
+```
+- Specify a custom remote base:
+```bash
+python3 vocabs_csv2jsonl.py --remote-base https://raw.githubusercontent.com/username/repo/main/docs/vocabularies
+```
+
+---
+
+## Notes
+- The script expects CSV files in `docs/vocabularies` named `<vocab>.csv` when using local mode or as a fallback.
+- Generated files are written under the destination directory; parent directories are created as needed.
+- CSVs must be encoded in UTF-8 (`utf-8-sig` supported) to handle optional BOMs.
